requirements.txt
python-telegram-bot[socks]
python-dotenv
ollama
